+++ lib/core/services/app_locator.dart
@@ -1,8 +1,9 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:get_it/get_it.dart';
 import 'package:http/http.dart' as http;
-import 'package:kwadra/features/auth/data/data_sources/kapi_auth_data_source.dart';
-import 'package:kwadra/features/auth/data/repositories/auth_repository_kapi.dart';
 
+import '../../features/auth/data/data_sources/firebase_auth_data_source.dart';
+import '../../features/auth/data/repositories/auth_repository_firebase.dart';
 import '../../features/auth/domain/repositories/auth_repository.dart';
 import '../../features/auth/domain/usecases/sign_in_with_email_and_password.dart';
 import '../../features/auth/presentation/bloc/auth_bloc.dart';
@@ -18,12 +19,14 @@ void setupLocator() {
 
   // Repositories
   locator.registerLazySingleton<AuthRepository>(
-      () => AuthRepositoryKwadraApi(locator()));
+      () => AuthRepositoryFirebase(locator()));
 
   // Data sources
-  locator.registerLazySingleton<KapiAuthDataSource>(
-      () => KapiAuthDataSourceImpl(client: locator()));
+  locator.registerLazySingleton<FirebaseAuthDataSource>(
+    () => FirebaseAuthDataSourceImpl(firebaseAuth: locator<FirebaseAuth>()),
+  );
 
   // External
   locator.registerLazySingleton(() => http.Client());
+  locator.registerLazySingleton<FirebaseAuth>(() => FirebaseAuth.instance);
 }