+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
@@ -8,9 +10,22 @@ import 'core/services/app_bloc_observer.dart';
 import 'core/services/app_locator.dart';
 import 'core/services/app_router.dart';
 import 'features/auth/presentation/bloc/auth_bloc.dart';
+import 'firebase_options.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
+  try {
+    await Firebase.initializeApp(
+      options: DefaultFirebaseOptions.currentPlatform,
+    );
+    if (kDebugMode) {
+      print("Firebase initialized successfully");
+    }
+  } catch (e) {
+    if (kDebugMode) {
+      print("Firebase initialization failed: $e");
+    }
+  }
   await dotenv.load(fileName: ".env");
   setupLocator();
   Bloc.observer = const AppBlocObserver();